Update pharmacologicVtePpxAdministered to use long instead of int.

+++ emerge/core/src/main/java/com/datafascia/emerge/harms/vte/PharmacologicVtePpxAdministered.java
@@ -40,7 +40,7 @@ public class PharmacologicVtePpxAdministered {
       String medName = medication.getName();
       for (PharmacologicVtePpxTypeEnum vtePpx : PharmacologicVtePpxTypeEnum.values()) {
         DateTimeDt timeTaken = (DateTimeDt) administration.getEffectiveTime();
-        Integer period = HarmsLookups.efficacyList.get(medName);
+        Long period = HarmsLookups.efficacyList.get(medName);
         if (vtePpx.toString().equalsIgnoreCase(medName.replace(" ", "_"))
             && HarmsLookups.withinDrugPeriod(timeTaken.getValue(), period)) {
           administered = true;
